refactor: Separate main JS file into modules
The main JS file has been separated into individual modules to profit form the advantages of it.
