+++ index.html
@@ -125,7 +125,6 @@
     ></script>
 
     <script src="node_modules/bootstrap/dist/js/bootstrap.min.js"></script>
-    <script src="src/index.js"></script>
-    <script src="src/unit-tests.js"></script>
+    <script src="src/index.js" type="module"></script>
   </body>
 </html>
